Add basic Resource Page from resources.dart branch
Moved the `resources.dart` to the `lib/` folder.
Added a ResourcesPage class that is a widget. Added that class to the
`main.dart` file to return resources page.
Added Resources button to the custom nav bar.

Resources page has a basic AppBar so it will still have a back button
after the Back Button issue is fixed.

Removed flutter-plugins files.